% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arkdb.R
\docType{package}
\name{arkdb-package}
\alias{arkdb}
\alias{arkdb-package}
\title{arkdb: Archive and Unarchive Databases Using Flat Files}
\description{
Flat text files provide a more robust, compressible,
and portable way to store tables.  This package provides convenient
functions for exporting tables from relational database connections
into compressed text files and streaming those text files back into
a database without requiring the whole table to fit in working memory.
}
\details{
It has two functions:
\itemize{
\item \code{\link[=ark]{ark()}}: archive a database into flat files, chunk by chunk.
\item \code{\link[=unark]{unark()}}: Unarchive flat files back int a database connection.
}

arkdb will work with any \code{DBI} supported connection.  This makes it
a convenient and robust way to migrate between different databases
as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/arkdb}
  \item Report bugs at \url{https://github.com/ropensci/arkdb/issues}
}

}
\author{
\strong{Maintainer}: Carl Boettiger \email{cboettig@gmail.com} (\href{https://orcid.org/0000-0002-1642-628X}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Richard FitzJohn [contributor]
}

}
