% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_db.R
\name{arkdb_delete_db}
\alias{arkdb_delete_db}
\title{delete the local arkdb database}
\usage{
arkdb_delete_db(db_dir = arkdb_dir(), ask = interactive())
}
\arguments{
\item{db_dir}{neon database location}

\item{ask}{Ask for confirmation first?}
}
\description{
delete the local arkdb database
}
\details{
Just a helper function that deletes the database
files.  Usually unnecessary but can be
helpful in resetting a corrupt database.
}
\examples{

# Create a db
dir <- tempfile()
db <- local_db(dir)

# Delete it
arkdb_delete_db(dir, ask = FALSE)


}
