% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_db.R
\name{local_db_disconnect}
\alias{local_db_disconnect}
\title{Disconnect from the arkdb database.}
\usage{
local_db_disconnect(db = local_db(), env = arkdb_cache)
}
\arguments{
\item{env}{The environment where the function looks for a connection.}
}
\description{
Disconnect from the arkdb database.
}
\details{
This function manually closes a connection to the \code{arkdb} database.
}
\examples{
\donttest{

## Disconnect from the database:
local_db_disconnect()

}
}
