% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamable_table.R
\name{streamable_base_csv}
\alias{streamable_base_csv}
\title{streamable csv using base R functions}
\usage{
streamable_base_csv()
}
\value{
a \code{streamable_table} object (S3)
}
\description{
streamable csv using base R functions
}
\details{
Follows the comma-separate-values standard using \code{\link[utils:read.table]{utils::read.table()}}
}
\seealso{
\code{\link[utils:read.table]{utils::read.table()}}, \code{\link[utils:write.table]{utils::write.table()}}
}
