% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamable_table.R
\name{streamable_parquet}
\alias{streamable_parquet}
\title{streamable chunked parquet using \code{arrow}}
\usage{
streamable_parquet()
}
\value{
a \code{streamable_table} object (S3)
}
\description{
streamable chunked parquet using \code{arrow}
}
\seealso{
\code{\link[arrow:read_parquet]{arrow::read_parquet()}}, \code{\link[arrow:write_parquet]{arrow::write_parquet()}}
}
