% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamable_table.R
\name{streamable_readr_tsv}
\alias{streamable_readr_tsv}
\title{streamable tsv using \code{readr}}
\usage{
streamable_readr_tsv()
}
\value{
a \code{streamable_table} object (S3)
}
\description{
streamable tsv using \code{readr}
}
\seealso{
\code{\link[readr:read_delim]{readr::read_tsv()}}, \code{\link[readr:write_delim]{readr::write_tsv()}}
}
