% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_db.R
\name{td_disconnect}
\alias{td_disconnect}
\title{Disconnect from the arkdb database.}
\usage{
td_disconnect(env = arkdb_cache)
}
\arguments{
\item{env}{The environment where the function looks for a connection.}
}
\description{
Disconnect from the arkdb database.
}
\details{
This function manually closes a connection to the \code{arkdb} database.
}
\examples{
\donttest{

## Disconnect from the database:
td_disconnect()

}
}
