% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unark.R
\name{unark}
\alias{unark}
\title{Unarchive a list of \code{*.tsv.bz2} files into a databaase}
\usage{
unark(files, lines = 10000L, drv = RSQLite::SQLite(), ...)
}
\arguments{
\item{files}{vector of filenames to be read in. Must be \code{*.tsv.bz2} format at present.}

\item{lines}{number of lines to read in a chunk.}

\item{drv}{an object that inherits from \linkS4class{DBIDriver},
or an existing \linkS4class{DBIConnection}
object (in order to clone an existing connection).}

\item{...}{authentication arguments needed by the DBMS instance; these
typically include \code{user}, \code{password}, \code{host}, \code{port}, \code{dbname}, etc.
For details see the appropriate \code{DBIDriver}.}
}
\value{
a database connection (invisibly)
}
\description{
Unarchive a list of \code{*.tsv.bz2} files into a databaase
}
\examples{
\donttest{

## Setup: create an archive.
sql_path <- tempdir()
db <- dbplyr::nycflights13_sqlite(sql_path)
dir <- file.path(tempdir(), "nycflights")
dir.create(dir)
ark(db, dir)
rm(db) # delete the original database, since we have an archive

## list all files in archive (full paths)
files <- list.files(dir, "[.tsv.gz]", full.names = TRUE)

## Read archived files into a new database (defaults to sqlite)
new_db <-  unark(files, dbname = "local.sqlite")

## Prove table is returned successfully.
library(dplyr)
tbl(new_db, "flights")

}
}
