% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamable_table.R
\name{streamable_table}
\alias{streamable_table}
\title{streamable table}
\usage{
streamable_table(read, write, extension)
}
\arguments{
\item{read}{read function. Arguments should be "\code{file}"
(must be able to take a \code{\link[=connection]{connection()}} object) and "\code{...}" (for)
additional arguments.}

\item{write}{write function. Arguments should be "\code{data}" (a data.frame),
\code{file} (must be able to take a \code{\link[=connection]{connection()}} object), and "\code{append}" logical,
append to existing file or create new?}

\item{extension}{file extension to use (e.g. "tsv", "csv")}
}
\value{
a \code{streamable_table} object (S3)
}
\description{
streamable table
}
\examples{

streamable_readr_tsv <- function() {
  streamable_table(
    function(file, ...) readr::read_tsv(file, ...),
    function(x, path, append)
      readr::write_tsv(x = x, path = path, append = append),
    "tsv")
}

}
