% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unark.R
\name{unark}
\alias{unark}
\title{Unarchive a list of \code{*.tsv.bz2} files into a databaase}
\usage{
unark(files, lines = 10000L, drv = RSQLite::SQLite(), ...)
}
\arguments{
\item{files}{vector of filenames to be read in. Must be \code{*.tsv.bz2} format at present.}

\item{lines}{number of lines to read in a chunk.}

\item{drv}{an object that inherits from \linkS4class{DBIDriver},
or an existing \linkS4class{DBIConnection}
object (in order to clone an existing connection).}

\item{...}{authentication arguments needed by the DBMS instance; these
typically include \code{user}, \code{password}, \code{host}, \code{port}, \code{dbname}, etc.
For details see the appropriate \code{DBIDriver}.}
}
\value{
a database connection (invisibly)
}
\description{
Unarchive a list of \code{*.tsv.bz2} files into a databaase
}
